requirements.txt
pyspark
sqlalchemy
psycopg2-binary
python-dotenv
pandas
